% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_nv.R
\name{check_nv}
\alias{check_nv}
\title{Check normal distribution of Likert-scale type item}
\usage{
check_nv(
  var,
  data,
  outliers = FALSE,
  ID = NULL,
  exclude_outliers = FALSE,
  plot = TRUE,
  title = NULL,
  ...
)
}
\arguments{
\item{var}{Variable to check, supplied as a string}

\item{data}{Dataframe from which the variable stems}

\item{outliers}{Whether function should check for outliers. If true, will return outlier values.}

\item{ID}{Column by which outliers should be identified. Function will return a vector consisting of the IDs of the outliers.}

\item{exclude_outliers}{Whether outliers should be removed prior to plotting the distribution of the variable.}

\item{plot}{Whether ND should be plotted}

\item{title}{Title. Used for plotting as well as for printing out test statistics. Defaults to variable name if not set.}

\item{...}{Arguments to pass to `plot_dist` for plotting of distribution}
}
\description{
Check normal distribution of variables detecting whether there are
outliers present and plotting the empirical distribution.
}
