% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_significance_stars.R
\name{get_significance_stars}
\alias{get_significance_stars}
\title{Generate significance signs from p values

Generate a vector of significance stars for a
set of p values.}
\usage{
get_significance_stars(
  p_vals,
  sign_levels = list(`.001` = "***", `.01` = "**", `.05` = "*")
)
}
\arguments{
\item{p_vals}{A vector of p values}

\item{sign_levels}{A list consisting of the significance levels as names and
significance signs as values. Names must be provided as strings and can have
a leading zero (but don't have to), i.e. can be in form of e.g. ".001" or "0.001".
Significance signs must also be provided as strings.}
}
\value{
A vector of significance signs corresponding to the p values
}
\description{
Generate significance signs from p values

Generate a vector of significance stars for a
set of p values.
}
