% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hayes_plot.R
\name{hayes_plot_model_1}
\alias{hayes_plot_model_1}
\title{Plot interactions generated using Process Hayes.}
\usage{
hayes_plot_model_1(
  hayes_data,
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  legend_title = NULL,
  SD = F,
  colorblind = F,
  APA = F
)
}
\arguments{
\item{hayes_data}{An object of class HayesData, containing the relevant information extracted from
the Process Hayes output}

\item{title}{Title of the plot}

\item{xlab}{X axis label}

\item{ylab}{Y axis label}

\item{legend_title}{Title above the legend}

\item{SD}{whether the labels of the legend should be replaced with +/-1 SD}

\item{colorblind}{Whether colorblind-safe colors should be used. If true, uses colors from the colorblind_1 palette in this package.}

\item{APA}{Whether black and white colors should be used.}
}
\description{
A function to visualize the interactions provided
by a the Process Hayes macro. Only works with hayesData objects,
which are the data class produced by `read_hayes_plot_data`.
}
