% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{palettes}
\alias{palettes}
\title{Colour palettes I find useful.}
\format{
An object of class \code{list} of length 1.
}
\usage{
palettes
}
\description{
Includes various helpful colour palettes, e.g. for creating plots with colourblind-safe colours.
}
\keyword{datasets}
