% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist.R
\name{plot_dist}
\alias{plot_dist}
\title{Plot distribution of Likert-scale type item}
\usage{
plot_dist(var, data, max_val = NULL, title = "", colorblind = F)
}
\arguments{
\item{var}{Variable to plot}

\item{data}{data.frame in which variable is contained}

\item{max_val}{Maximum value of the item being analyzed (optional). If not supplied, the x axis will scale automatically. If supplied,
the x axis will scale between 0 and max_val.}

\item{title}{Title for the plot, optional}

\item{colorblind}{Optional change of colors using the colorblind_1 palette from this package.}
}
\value{
Plot of response pattern for the item
}
\description{
Plot distribution of Likert-scale type item
}
