% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hayes_plot_data_R.R
\name{read_hayes_plot_data_R}
\alias{read_hayes_plot_data_R}
\title{Read Hayes data as .txt file into R}
\usage{
read_hayes_plot_data_R(data)
}
\arguments{
\item{data}{The output of a process function call in R, captured using capture.output() from utils (or something else with an
equivalent function).}
}
\value{
A list of class HayesData consisting of the plot_data for visualizing the interaction as well as the various significance 
measures from the Hayes SPSS output as well as the model description.
}
\description{
The process Hayes Macro is quite nice for a lot of different analysis types. However,
the R implementation is not quite there yet in terms of usability, which is where this function comes in. You
can capture the output of the R implementation of Process Hayes using the capture.output() function from 
the utils package to save the output of the process function. You can then pass the captured output into 
this function to get data that is somewhat more usable, e.g. for plotting interactions or accessing certain
coefficients quickly.

To extract the data from the Hayes output, this function uses the text in specific lines of the Hayes module. 
It might be possible that these change between versions. The current version this function works with Hayes
version 4.0.
}
