% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hayes_plot_data_spss.R
\name{read_hayes_plot_data_spss}
\alias{read_hayes_plot_data_spss}
\title{Read Hayes data as .txt file into R}
\usage{
read_hayes_plot_data_spss(path)
}
\arguments{
\item{path}{Path of the .txt file with the SPSS output script.}
}
\value{
A list of class HayesData consisting of the plot_data for visualizing the interaction as well as the various significance 
measures from the Hayes SPSS output as well as the model description.
}
\description{
The process Hayes Macro is quite nice for a lot of different analysis types. However,
the R implementation is not quite there yet in terms of usability, whereas the plots you can generate
with SPSS are straight up not pretty. This function thus reads in the SPSS output of a Process Hayes 
model (provided as a .txt file), finds the plot data and converts it into a R dataframe. It also stores additional
parameters of interest.

To extract the data from the Hayes output, this function uses the text in specific lines of the Hayes module. 
It might be possible that these change between versions. The current version fo this function works with Hayes
version 4.0.

There will be a similar method for R methods called read_hayes_plot_data_R sometime.
}
