% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_nv.R
\name{check_nv}
\alias{check_nv}
\title{Check normal distribution of Likert-scale type item}
\usage{
check_nv(var, dat, outliers = FALSE, ID = NULL, plot = T)
}
\arguments{
\item{var}{Variable to check, supplied as a string}

\item{dat}{Dataframe from which the variable stems}

\item{outliers}{Whether function should return the outliers}

\item{ID}{Column by which outliers should be identified}

\item{plot}{Whether ND should be plotted}
}
\description{
Check normal distribution of variables detecting whether there are
outliers present and plotting the empirical distribution.
}
