% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nv.R
\name{plot_dist}
\alias{plot_dist}
\title{Plot distribution of Likert-scale type item}
\usage{
plot_dist(var, dat, title = "")
}
\arguments{
\item{var}{Variable to plot}

\item{dat}{data.frame in which variable is contained}

\item{title}{Title for the plot, optional}
}
\value{
Plot of response pattern for the item
}
\description{
Plot distribution of Likert-scale type item
}
