% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist.R
\name{plot_dist}
\alias{plot_dist}
\title{Plot distribution of Likert-scale type item}
\usage{
plot_dist(
  var,
  data,
  max_val = NULL,
  title = "",
  colorblind = F,
  density = T,
  binwidth = 0.5,
  return = F
)
}
\arguments{
\item{var}{Variable to plot}

\item{data}{data.frame in which variable is contained}

\item{max_val}{Maximum value of the item being analyzed (optional). If not supplied, the x axis will scale automatically. If supplied,
the x axis will scale between 0 and max_val.}

\item{title}{Title for the plot, optional}

\item{colorblind}{Optional change of colors using the colorblind_1 palette from this package.}

\item{density}{Whether density curve should be plotted. Helpful to deactivate for single-item analysis.}

\item{binwidth}{Width of bins. Should be left at 0.5 for Likert-Scale items (aggregated) or set to 1 for single-item analysis.}

\item{return}{If set to true, returns the plot instead of printing it. Use if you want to arrange multiple plots e.g. using `ggarrange`.}
}
\value{
Plot of response pattern for the item
}
\description{
Plot distribution of Likert-scale type item
}
